Getting this pipeline off the ground.
This is the start of the piepline. As of now, it takes in nifti images,
optionally with label maps, and computes a suite of GLCM features from
them. Label maps can have multiple values. I'm currently working on
truncating ROIs received from label maps into a smaller arrays so that
feature calculators don't have to iterate over zero values, and failing
multiple times; this is the next step.