Added morphology and Statistic module.
A variety of changes come in this update. After struggling with
correctly calculating surface area, I opted for a kernel-convolution
method. It over-estimates SA, but probably not by too much. Most of the
rest of the morphology features were taken from Slicer's Heterogeneity
CAD module. I bet there are some more interesting ones to include. The
statistics module was also tacked on; it takes first-order statistics of
the flattened array of pixels. The general infrastructure of the program
has improved, and modules can now be added on regularly. TO-DO: Test on
more real-world data. I currently have not dealt with the problem of
label-maps with different dimensions from their target volume, which
happens often with DICOM-seg objects. It may be a nice future project to
borrow a python-to-DICOM reader, although the business of maintaing a
DICOM-reading program may not be worth getting into. Add, maybe: Gabor
filters, Local binary filters, Reisz filters and other 3D wavelets
transforms, run-length features (they performed well in the past!)